requirements.txt
openai
pillow
torchvision
numpy
torch